feat: implement review screen and refactor shared composables

+++ data/src/main/java/com/london/data/datasource/remote/reviews/model/ReviewResponse.kt
@@ -30,5 +30,5 @@ data class AuthorDetailsResponse(
     @SerialName("avatar_path")
     val authorPictureUrl: String?,
     @SerialName("rating")
-    val rating: Float?
+    val rating: Double?
 )
