+++ domain/src/main/java/com/london/domain/entity/review/ReviewEntity.kt
@@ -14,5 +14,5 @@ data class AuthorDetails(
     val name: String,
     val username: String,
     val profileUrl: String,
-    val rating: Float
+    val rating: Double
 )
